chore(tabGroup): move tabGroup component related code to components/contentful folder
