+++ frontend/apps/marketing/src/components/contentful/tabGroup/TabGroupContentfulDefinition.ts
