+++ frontend/apps/marketing/src/components/contentful/tabGroup/index.ts
